pyproject.toml
[project]
name = "sixty-nuts"
version = "0.0.6"
description = "Sixty Nuts - NIP-60 Cashu Wallet Implementation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "shroominic", email = "contact@shroominic.com" }]
keywords = [
    "cashu",
    "ecash",
    "nostr",
    "nip-60",
    "bitcoin",
    "lightning",
    "wallet",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security :: Cryptography",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "httpx>=0.28.1",
    "websockets>=15.0.1",
    "bech32>=1.2.0",
    "coincurve>=21.0.0",
    "cryptography>=45.0.3",
    "cbor2>=5.6.5",
    "python-dotenv>=1.1.0",
]
urls = { "Repository" = "https://github.com/shroominic/sixty-nuts" }

[dependency-groups]
dev = ["mypy>=1.16.0", "pytest-asyncio>=1.0.0", "ruff>=0.11.12"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
pythonpath = ["."]
markers = ["asyncio: mark test as async"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
